requirements.txt
Flask==2.0.1
python-dotenv==0.15.0
google-generativeai==0.7.1
gunicorn==20.1.0
